This commit includes the data as well
Success story contains the best build to date, which is 72.92% accuracy. Weights contains the weights of the last run. The neural network now only saves when a highscore has been reached and I am still conducting test to determine best learning pace. It seems to be maybe 0.1 but I still don't know.